* main: (107 commits)
  Updated Release Notes.
  A bit of cleanup for now, more to follow, but moving on to other roadblocks.
  Comment out debugging messages, will remove before final merge.  Clean up the logic, remove some rough edges.
  Add a filter to process options and report deprecation warning.
  Remove dangling define in netcdf_meta.h template file.
  No effect other than to remove a personal annoyance I introduced in the first place.
  Updated the nc-config script with the following:
  adding deprecation error for usage of NETCDF_ENABLE_NETCDF4
  removing cmake alias variable for netcdf4
  removing script
  replacing something that removes semicolons
  another one
  fixing some that were missing?
  removing variables
  Replacing NC_USE_STATIC_CRT with NETCDF_USE_STATIC_CRT
  Replace NC_FIND_SHARED_LIBS with NETCDF_FIND_SHARED_LIBS
  Replace ENABLE_XGETOPT with NETCDF_ENABLE_XGETOPT
  Replace ENABLE_UNIT_TESTS with NETCDF_ENABLE_UNIT_TESTS
  Replace ENABLE_TESTS with NETCDF_ENABLE_TESTS
  Replace ENABLE_STRICT_NULL_BYTE_HEADER_PADDING with NETCDF_ENABLE_STRICT_NULL_BYTE_HEADER_PADDING
  ...

+++ nc_test4/tst_camrun.c
@@ -684,7 +684,7 @@ get_mem_used2(int *mem_used)
    {
       (void)fscanf(pf, "%u %u %u %u %u %u", &size, &resident, &share,
 	     &text, &lib, &data);
-      *mem_used = (data * page_size) / MEGABYTE;
+      *mem_used = (int)(data * page_size) / MEGABYTE;
    }
    else
       *mem_used = -1;
